Add a new coat of paint, add variations extension.
The sphinx-rtd-theme has been modified to make it look more Panda-ish;
this mainly entails putting the logo up top and changing the color
scheme. tcdude's CSS overrides have been integrated as well.

Additionally, the sphinx-variations extension has been added to support
building both the Python and C++ versions of the manual. Links for
switching have been added to the top of the sidebar as well.